create useChat hook
